DMSafeKVC: Add KeysAffectingSuperclassOf() macro
This addresses a deficiency in the frameworks when overriding a method
whose value has dependencies.

Example:

  @implementation Derived
  - (NSString *)title {
      return [NSString stringWithFormat:@"%@ (%ld)",
	          [super title], self.count];
  }
  + (NSSet *)keyPathsForValuesAffectingTitle {
      return [KeysAffectingSuperclassOf([Derived class])
	          setByAddingObject:K(count)];
  }
  @end

Calling KeysAffectingSuperclassOf() takes the class of the
implementation uses _cmd and calls +keyPathsForValuesAffectingTitle on
the superclass of the given class. If the superclass doesn't declare
that method or returns nil, KeysAffectingSuperclassOf() returns an empty
set so the -setByAddingObject: idiom still works.

The class argument must (unfortunately) be explicit, not 'self'. This is
necessary for further subclasses not to crash. (A additional subclass
with no extra methods will infinitely loop, as Derived will now be the
superclass.) I wish there were a way to refer to the current class in a
macro, but alas.

Note that this calls the +keyPathsForValuesAffecting... methods
directly, instead of via +keyPathsForValuesAffectingValueForKey:,
because sending that to super would redirect the call to the method on
the actual class, not the superclass. Accordingly, dependencies
specified that way will NOT be picked up by this code (caveat!).